JJ dev 05 - Fixed Transform/Lock Issue, Kicking, and added many goodies. (#46)
* 2nd Attempt. Reaverbot Bomb Update

Added to the gitignore file to ignore the Intermediate and Saved folders. Added the bomb once again to get it started in the scene. Mostly I think things will look much better when we get some real lighting in there.

* Update M_ReaverbotBomb.uasset

* Cleanup

* Nothing Crazy. Fixed the kick issue, but the BusterGunTransform Issue is still present.

Currently, when the player transforms to the buster gun, the gun automatically points to the camera direction. This is not the correct behavior. Still searching for the solution.

* nvm. Finally Found it. It was in the anim BP and a single boolean lol.

* Squashed Bugs. Added some Goodies!

Fixed the issue with the targeting/Transformation Buster Gun. Now it behaves closers to MML. Also did some other minor fixes/tweaks to the Kick. I also Migrated some Stylized VFX that I've purchased for this project that we can customize for Megaman's various weapons and plasma. Should make for a nice addition and give it some production value.

Let me know if you see all these changes.