Major Zone 3 Structure Changes
The environment now plays out more like how it would if we were in the physical space. The tower has been dramatically extended above a water line, and the relevant assets have been updated to reflect this change. We can now start building final assets based loosely on this blockout.